% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_forest_class.R
\name{export_forest_plots-methods}
\alias{export_forest_plots-methods}
\alias{export_forest_plots,GWASForestPlots-method}
\alias{export_forest_plots,MRForestPlots-method}
\title{S4 methods for export_forest_plots}
\description{
Internal aliases used to document the concrete S4 methods for
\code{export_forest_plots}. Users should normally consult
\code{?export_forest_plots}.
}
\keyword{internal}
