% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_scatter_class.R
\name{export_scatter_plots-methods}
\alias{export_scatter_plots-methods}
\alias{export_scatter_plots,MRScatterPlots-method}
\title{S4 methods for export_scatter_plots}
\description{
Internal aliases used to document the concrete S4 method for
\code{export_scatter_plots}. Users should normally consult
\code{?export_scatter_plots}.
}
\keyword{internal}
