% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_49item.R
\name{fi_49item}
\alias{fi_49item}
\title{Frailty Index Dataset (49-items)}
\format{
A data frame.
}
\description{
A dataset containing variables used to compute a 49-item frailty index (FI).
The exact variables/columns depend on your project definitions.
}
\keyword{datasets}
