% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fried_frailty.R
\name{fried_frailty}
\alias{fried_frailty}
\title{Fried Frailty Phenotype Dataset}
\format{
A data frame.
}
\description{
A dataset containing variables used to compute Fried frailty phenotype measures.
The exact variables/columns depend on your project definitions.
}
\keyword{datasets}
