% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_transform.R
\name{harmonize_mr_data}
\alias{harmonize_mr_data}
\title{Harmonize exposure and outcome SNP data}
\usage{
harmonize_mr_data(df, Outcome = NULL, Exposure = NULL)
}
\arguments{
\item{df}{Data frame containing required columns (beta_outcome, se_outcome,
beta_exposure, se_exposure) and metadata (Outcome, Exposure).}

\item{Outcome}{Optional; Character string to set as the Outcome name.}

\item{Exposure}{Optional; Character string to set as the Exposure name.}
}
\value{
A harmonized data frame with 'Instrument' and allele columns.
}
\description{
Harmonizes alleles and ensures consistent effect direction between exposure
and outcome datasets. Standardizes exposure effects to be positive.
}
\examples{
input_test3 <- harmonize_mr_data(df = fi_49item)
head(input_test3)
}
