% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_data.R
\name{merged_data}
\alias{merged_data}
\title{Merged Frailty Index Dataset (49-items and fried frailty)}
\format{
A data frame.
}
\description{
A dataset containing variables used to compute a 49-item frailty index (FI) and Fried frailty.
The exact variables/columns depend on your project definitions.
}
\keyword{datasets}
