% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_scatter_class.R
\name{showplot-methods}
\alias{showplot-methods}
\alias{showplot,MRScatterPlots-method}
\alias{showplot,GWASForestPlots-method}
\alias{showplot,MRForestPlots-method}
\title{S4 methods for showplot}
\description{
Internal documentation entries for S4 methods. These aliases satisfy
R CMD check without creating user-facing method help pages.
}
\keyword{internal}
