% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_transform.R
\name{format_mr_input}
\alias{format_mr_input}
\title{Format vectors into a Mendelian Randomization input data frame}
\usage{
format_mr_input(
  Instrument,
  beta_exposure,
  se_exposure,
  beta_outcome,
  se_outcome,
  Outcome,
  Exposure,
  ALLELE1 = NULL,
  ALLELE0 = NULL,
  A1FREQ = NULL
)
}
\arguments{
\item{Instrument}{Character vector of instrument/SNP identifiers.}

\item{beta_exposure}{Numeric vector of exposure effects.}

\item{se_exposure}{Numeric vector of exposure standard errors.}

\item{beta_outcome}{Numeric vector of outcome effects.}

\item{se_outcome}{Numeric vector of outcome standard errors.}

\item{Outcome}{Character string or vector for outcome names (Mandatory).}

\item{Exposure}{Character string or vector for exposure names (Mandatory).}

\item{ALLELE1}{Optional character vector for non-effect alleles.}

\item{ALLELE0}{Optional character vector for effect alleles.}

\item{A1FREQ}{Optional numeric vector for effect allele frequencies.}
}
\value{
A data frame formatted for MR analysis.
}
\description{
Format vectors into a Mendelian Randomization input data frame
}
\examples{
data("fi_49item")

input_test1 <- format_mr_input(
  Instrument = fi_49item$Instrument,
  beta_exposure = fi_49item$beta_exposure,
  se_exposure = fi_49item$se_exposure,
  beta_outcome = fi_49item$beta_outcome,
  se_outcome = fi_49item$se_outcome,
  Outcome = fi_49item$Outcome,
  Exposure = fi_49item$Exposure
)
head(input_test1)

input_test2 <- format_mr_input(
  Instrument = fi_49item$Instrument,
  beta_exposure = fi_49item$beta_exposure,
  se_exposure = fi_49item$se_exposure,
  beta_outcome = fi_49item$beta_outcome,
  se_outcome = fi_49item$se_outcome,
  Outcome = fi_49item$Outcome,
  Exposure = fi_49item$Exposure,
  ALLELE1 = fi_49item$NEA_outcome,
  ALLELE0 = fi_49item$EA_outcome,
  A1FREQ = fi_49item$A1FREQ_outcome
)
head(input_test2)
}
